+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add *GLIRenderer*.
+- Add *GLITransform*.
+- Add `width` and `height` in *GLIRenderTarget*.
+
 ## [0.1.0] - 2019-05-31
 
 ### Added